Move coordinate classes to static_layout_provider
